WIP make lr work
