+++ lr_trie/src/db.rs
@@ -9,7 +9,7 @@ use crate::error::MemDBError;
 /// "DB" defines the "trait" of trie and database interaction.
 /// You should first write the data to the cache and write the data
 /// to the database in bulk after the end of a set of operations.
-pub trait Database: Send + Sync + Clone + Default {
+pub trait Database: Send + Sync + Clone + Default + std::fmt::Debug {
     type Error: Error;
 
     fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;
