+++ lr_trie/src/result.rs
@@ -1,0 +1,4 @@
+use crate::error::TrieError;
+use std::result::Result as StdResult;
+
+pub type Result<T> = StdResult<T, TrieError>;
